+++ changelog.md
@@ -3,6 +3,7 @@
 - @blackheaven
   - fix: remove `putMapping` deprecation
   - strong type backend
+  - add dynamic backend operations
 
 0.23.0.1
 ========